+++ nix/configuration.nix
@@ -1,0 +1,26 @@
+{ config, lib, pkgs, ... }:
+
+{
+  virtualisation.vmVariant = {
+    virtualisation = {
+      memorySize = 2048; # Use 2048MiB memory.
+      cores = 3;
+    };
+  };
+
+  services.xserver.enable = true;
+  services.xserver.displayManager.gdm.enable = true;
+  services.xserver.desktopManager.gnome.enable = true;
+
+  users.users.alice = {
+    isNormalUser = true;
+    extraGroups = [ "wheel" ];
+    packages = with pkgs; [
+      firefox
+    ];
+    password = "genealogos";
+  };
+
+  services.genealogos.enable = true;
+}
+