+++ nix/overlays.nix
@@ -1,0 +1,4 @@
+{ pkgs, packages }:
+_: _: {
+  genealogos-api = packages.genealogos-api;
+}